Remove duplicate documentation
Having moved the documentation into its own file in the doc/ folder, it
is no longer necessary as part of the plugin text.  Indeed, leaving it
there introduces the danger that documentation might get updated in one
place and not in another, leading to uncertainty about which is correct.

Remove documentation comments from the plugin, leaving the new version
inside the doc/ folder as the canonical version.