Fixed wincmd to work properly when using CalendarH
When using :CalendarH, if you press [enter] on a date string, a diary file will be opened in the same window. Thus a calendar view will disappear. This is not the action we want, so I changed the script to open a diary file in a different window. To do this cleverer, you can use "wincmd l" in :Calendar mode and "wincmd k" in :CalendarH mode, but in most cases "wincmd w" (the command I chose) will do the right thing.